requirements.txt
pdfplumber==0.11.0
torch==2.1.2
paddle==2.6.0
paddleocr==2.7.0.0
langchain==0.0.352
loguru==0.7.2
opencv-python==4.6.0.66
unstructured==0.12.6
spire.doc==12.7.1
pyOpenSSL==24.2.1
transformers==4.40.0
typing=3.6.6
fitz=0.0.1.dev2
fastapi=0.111.0
requests=2.30.0

